# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/2001adarsh/Introduction-to-Microservices/handlers
  Product:
    description: Product defines the structure of an API Product
    properties:
      description:
        description: description of the product
        type: string
        x-go-name: Description
      id:
        description: id of the product
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: name of the product
        type: string
        x-go-name: Name
      price:
        description: cost of the product
        format: float
        minimum: 0
        type: number
        x-go-name: Price
      sku:
        description: SKU of the product
        example: abc-bcd-def
        type: string
        x-go-name: SKU
    required:
    - name
    - sku
    type: object
    x-go-package: github.com/2001adarsh/Introduction-to-Microservices/data
  ValidationError:
    description: ValidationError is a collection of validation error messages
    properties:
      messages:
        items:
          type: string
        type: array
        x-go-name: Messages
    type: object
    x-go-package: github.com/2001adarsh/Introduction-to-Microservices/handlers
host: localhost
info:
  description: Documentation for Product API
  title: of Product API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns all products in DB as a list
      operationId: listAllProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
        "500":
          $ref: '#/responses/internalServerError'
      tags:
      - products
    post:
      description: Create a new product
      operationId: createProduct
      parameters:
      - description: |-
          Product data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    put:
      description: Update a products details
      operationId: updateProduct
      parameters:
      - description: |-
          Product data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "201":
          $ref: '#/responses/noContent'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - products
  /products/{id}:
    delete:
      description: deletes a product from the database
      operationId: deleteProduct
      parameters:
      - description: The ID of the product from database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContent'
        "404":
          $ref: '#/responses/errorResponse'
        "500":
          $ref: '#/responses/internalServerError'
      tags:
      - products
    get:
      description: Return a list of products from the database
      operationId: singleProduct
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "500":
          $ref: '#/responses/internalServerError'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  internalServerError:
    description: ""
  noContent:
    description: ""
  productResponse:
    description: Data structure representing a single product
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of Products returns in response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"